Added a patch from Joshua Kinard <kumba@gentoo.org> to clean up the Pegasos hacks in genkernel.  This is from bug #193826.  This is genkernel 3.4.9_pre7 for testing.
git-svn-id: svn+ssh://svn.gentoo.org/var/svnroot/genkernel/trunk@554 67a159dc-881f-0410-a524-ba9dfbe2cb84

+++ gen_initramfs.sh
@@ -475,21 +475,21 @@ create_initramfs() {
 	gzip -9 "${CPIO}"
 	mv -f "${CPIO}.gz" "${CPIO}"
 
-	# Pegasos hack for merging the initramfs into the kernel at compile time
-	[ "${KERNEL_MAKE_DIRECTIVE}" == 'zImage.initrd' -a "${GENERATE_Z_IMAGE}" = '1' ] ||
-		[ "${KERNEL_MAKE_DIRECTIVE_2}" == 'zImage.initrd' -a "${GENERATE_Z_IMAGE}" = '1' ] &&
+	if [ "${ENABLE_PEGASOS_HACKS}" = 'yes' ]
+	then
+			# Pegasos hack for merging the initramfs into the kernel at compile time
 			cp ${TMPDIR}/initramfs-${KV} ${KERNEL_DIR}/arch/powerpc/boot/ramdisk.image.gz &&
 			rm ${TMPDIR}/initramfs-${KV}
-
-	# Mips also mimics Pegasos to merge the initramfs into the kernel
-	if [ ${BUILD_INITRAMFS} -eq 1 ]; then
+	elif [ ${BUILD_INITRAMFS} -eq 1 ]
+	then
+		# Mips also mimics Pegasos to merge the initramfs into the kernel
 		cp ${TMPDIR}/initramfs-${KV} ${KERNEL_DIR}/initramfs.cpio.gz
 		gunzip -f ${KERNEL_DIR}/initramfs.cpio.gz
 	fi
 
 	if ! isTrue "${CMD_NOINSTALL}"
 	then
-		if [ "${GENERATE_Z_IMAGE}" != '1' ]
+		if [ "${ENABLE_PEGASOS_HACKS}" != 'yes' ]
 		then
 			copy_image_with_preserve "initramfs" \
 				"${TMPDIR}/initramfs-${KV}" \
